Analysis of neutrons data measured on 2019-02-14.
Improved the functions creating the datasets and parameters for batch fitting.

Created two additional functions from previously written pieces of code, in view of the need to reuse them multiple times:
* one that computes the best fit parameters from the Minimizer.Result object created during the fitting process;
* one that plots the data + fitting curves with best fit parameters.

Re-ran the section for finding the best fit parameters shared among all datasets for the (10 10 0) peak.
The result is satisfying now.